Add grafana persist option to istio helm chart
On pod death, the current grafana chart does not persist any
configuration changes made in the UI (like adding alerts).
This adds a boolean option `grafana.persist` to create a
`PersistentVolumeClaim` that mounts to the existing grafana
data volume.

There is an additional option for `storageClassName` to override the
default storage class.